Improved :keyword-typos linter implementation by reading the forms again
This avoids extra occurrences of keywords like :line :column etc. that
occur if we use the forms as read during the analysis phase.  It also
corrects a previously experienced problem with using this approach
where the reader would throw an exception if it read ::ns-alias/name
keywords, which occur in several crucible projects.  Nicola Mometto
gave the idea for the current solution, which is to bind *ns* while
re-reading the forms.