Recognize and print a special message for a new tools.analyzer exception
tools.analyzer already threw an exception when it found a form like
(. object (something here) more-here) because it treated the
'more-here' as erroneous.  Clojure 1.5.1 ignores the 'more-here'
silently.

Nicola Mometto added a more informative exception for this case to
tools.analyzer.  This commit adds code to check for that exception and
print a brief informative message to the developer.  It needs the TBD
replaced with a correct example or examples.