Add test that fails unless in-ns switching works
This new test is patterned after some code found in namespace
incanter.sql-tests, which caused an exception with Eastwood 0.1.1, but
no longer throws an exception due to some changes in the way Eastwood
calls tools.analyzer(.jvm) that were made for other reasons.

Add a test for this that is executed every time 'lein test' is run, so
we don't need to add Incanter to the crucible just to cover this case.