Copy most dependencies source files into Eastwood source
This is an attempt to allow Eastwood to lint more projects, even if
they depend upon a version of a project incompatible with the one that
Eastwood itself depends upon.  Now they will be in separate namespaces
(as long as no linted project uses eastwood as its first namespace
segment).