Improve the line/col numbers for :unlimited-use warnings
The previous 2 commits causes namespaces to no longer have :line
:column metadata associated with them, because they are created before
tools.reader reads them.

The new line/col data implemented with this commit are more precise
than the old ones anyway, because they are the location of the actual
namespace where it is :use'd, not of the namespace that :use's it.