Add back in several debug options that were recently removed
The code that adds back the debug capability of pprint'ing forms after
they are read, before and after macroexpansion, have been enhanced to
show shorter output.  It does this by removing the keys :column
:end-line :end-column and :file from metadata maps, which are added by
tools.reader.  The :line keys are left in, to indicate where in the
forms this kind of tools.reader metadata is present.