Remove obsolete sanity check
When I originally added this sanity check, fn-arglists (-> a
:arglists) and fn-arglists2 (-> a :init :arglists) were always the
same, but I didn't know it at the time and wanted to verify.

Somewhere between tools.analyzer(.jvm) versions 0.2.2 and 0.6.0, one
of these changed, so that fn-sigs and fn-sigs2 are frequently
different.  This appears to be completely intentional and expected, so
I am removing this sanity check and some expressions above it that
were only calculated for the purposes of performing the sanity check.