Remove warn-on-reflections pass from tools.analyzer(.jvm), since
at least by default in the crucible tests, I have set up the Leiningen
project.clj files to warn on reflection already, and having this pass
causes duplicates to be printed.

I may want to consider removing that from the Leiningen project.clj
files, and rely on the warn-on-reflection pass in tools.analyzer(.jvm)
instead.