Additions and refinements to :constant-test linter
It has been run on all current projects in crucible.  No crashes, and
most of the warnings look reasonable.  I disabled them for Carmine and
core.match because they have some macros and/or tests that cause
hundreds of :constant-test warnings to be produced.  Definitely would
be nice to have a way to disable the warning inside of particular
macro definitions, and have that disabling survive the macroexpansion
process.