Work on a branch towards making suspicious-expression based on asts only
This should eliminate at least some undesired :suspicious-expression
warning messages for forms like (-> 1 (= 1)).

I also believe that once all but a tiny fraction of linters are using
the asts and not the source code text or read-in forms, it will be
easier to have 1 way to suppress warnings on a per-expression basis
via source code annotations (e.g. in metadata).  Trying to devise an
implementation that works for linters that use text or read-in forms
as well sounds like a pain.