Add new option :disable-linter-name-checks
Only intended for my scripting purposes, so that in the future if I
add a new linter that is enabled by default, but I want to explicitly
disable it in the lint.sh script of some project in the crucible, I
can run that lint.sh script with the later and this version of
Eastwood, as long as I use :disable-lint-name-checks true.

Without this option, the older version of Eastwood that does not
implement the new linter will stop after printing an error message
about the unrecognized linter name.