:constant-test warning disabling based on macroexpansion context
Similar to support for :suspicious-expression warnings added recently,
and mechanism is nearly identical.

A difference: The :constant-test linter has many warnings for ASTs
where the enclosing-macros function returns an empty sequence of
enclosing macros.  This is probably due to some bug in how I fill in
the values for the :eastwood/ancestors key.