Fix issue with constant-test linter that was preventing disabling from working
This fix has the unfortunate side effect of making some of the
expressions shown in the warning messages more difficult to
understand.  I can work on improving that later.

Also add some more constant-test disabling to config.clj for Encore
and Timbre libraries, which make the constant-test linter much quieter
on a couple of projects in the crucible where it was way too noisy.