Enhance wrong-arity linter to catch wrong number of args in new case
Currently tools.analyzer.jvm always uses the :arglists metadata to
determine whether to mark and AST with a :maybe-wrong-arity key.  Thus
if someone were to override the :arglists metadata on a function in
such a way that a call with a wrong number of args appeared to have a
correct number of args, Eastwood would never have warned about it.

Now Eastwood ignores the :maybe-wrong-arity key in the ASTs, and does
its own independent wrong-arity checking, based upon :arglists
provided in its configs if they are there, otherwise the :arglists
metadata on the function.

Also added a couple more configs to disable unwanted wrong-arity
warnings in crucible projects.