Update some test cases so they continue to pass with Clojure 1.5.1
I don't know why I have not noticed these failures before.  Mostly
they are because of new macros added in Clojure 1.6.0 that are in some
of the test cases.  Other cases appear to be different seq order for
hash sets or hash maps due to changes in hash function in Clojure
1.6.0.