Actually use the linters that were loaded from files
As currently implemented, custom linters are always enabled, because
the keys of the custom linters map are added directly to the list of
linters. This works well when people are writing their own linters, so
they can comment out linters they don't want. It might be a problem if
linters are being distributed alongside libraries (or as a library!).