Make the custom linter example more realistic
Now uses tools.analyzer to analyzer the namespace for a simple issue -
function names that go against a style guide. In this case, pretend
the word 'custom' has been banned for being too generic.

Also remove the ns definition, because it's unnecessary - custom lint
files will be evaluated inside eastwood.util anyway. So we can make
use of the aliases and functions defined there in the example linter.