Add first draft of duplicate-params linter
It only checks fn arg vectors, including destructuring that might be
present there.

It also does some checks on :as and :or values of associative
destructuring forms, but again, only in fn arg vectors with this
commit.

It would be good to extend all of these to check all uses of
destructuring, e.g. in let, loop, if-let, etc. bindings.