Remove callbacks + cleanups
This commit removes most of the callback code used for reporting in
Eastwood, in favour of a set of multimethods found in the
`reporting-callbacks` ns

Also, state-handling has gone from atoms to a more functional
approach, and error handling is now largely done by exceptions rather
than passing maps with magic keys.

An effort has been done to remove the passing of `opts` and rather
extract the needed values from `opts` before calling the next function.