Cleanup test suite, dev repl, increase CI job parallelism (#361)
* CircleCI: use Lein 2.9.5

* Make dogfooding pass

* Make the test suite pass without -dev

And particularly, without the `leiningen-core` dependency which makes it plausible that the test suite could be passing for fragile reasons.

* Extract :warn-on-reflection profile

This way, one can opt to launch a repl process without `*warn-on-reflection*`, which can be noisy, in face of arbitrary deps outside our control.

* Move `leiningen.eastwood` to its own source path

This way one can use tools.namespace `(refresh)` without needing the leiningen-core dependency to be present (which tends to be a bad practice in non-plugins)

* project.clj `:eastwood`: remove `:only-modified true`

Because the Eastwood implementation can always change, caching linting results can easily cache past implementations' output, leading to false negatives.

* Make the matrix fully parallel