[Fix #341] Remove `^:const` metadata prior to processing a given form
`^:const` is not part of any public API described in https://github.com/clojure/clojure-site.

However usage of it can be still found in the wild, perhaps influenced by past knowledge.

https://github.com/jonase/eastwood/issues/341 showed that attempting to process a defn having ``^:const` would throw an exception.
This commit proposes removing that piece of metadata, since it doesn't alter semantics and saves the effort of investigating
a more intrincate integration with tools.analyzer (for what I believe is not a public API anyway).

It's worty noting that https://clojure.org/reference/compilation#directlinking **is** a public API,
and something that can be used in production environments instead of `^:const`, without fearing a clash with Eastwood whatsoever.

Fixes https://github.com/jonase/eastwood/issues/341