[Fix #339] Upgrade `linter-executor-atom` to `pmap`
 I've used `pmap` over the last 12 months, continuously (in CI, and work/personal projects) observing no issues.

So it seems a safe and convenient default to improve.

`clojure.core/pmap` is not too bad of a choice (vs. other pmap-like functions) because its implementation caps the maximum amount of spawned threads to a number relative to CPU count.

So the workloads, which are CPU-bound, will be treated as such, without possibly hogging the OS with an excess of threads.